% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgp1.R
\docType{package}
\name{bcgp1}
\alias{bcgp1}
\alias{bcgp1-package}
\title{bcgp1: a package for implementing Bayesian Composite Gaussian Processes}
\description{
The bcgp1 package performs MCMC and blah, blah, blah
}
\section{These are the functions}{

bcgp, MCMC, ...
}

