% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgpMCMC.R
\name{bcgpMCMC}
\alias{bcgpMCMC}
\title{Draw samples from a bcgp model}
\usage{
bcgpMCMC(xTrain, yTrain, priors, inits, numUpdates, numAdapt, burnin,
  nmcmc)
}
\arguments{
\item{yTrain}{Avector containing the observed response values in the training
set.}

\item{numUpdates}{The number of updates in the proposal stepsize adaptation phase.}

\item{numAdapt}{The number of samples within each update in the proposal stepsize
adaptation phase.}

\item{burnin}{The number of burnin samples to discard after the stepsize
adaptation phase is finished}

\item{X}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{prior}{A list containing the values for the prior parameters.}
}
\value{
An object of S4 class \code{bcgp} representing the fitted results.
}
\description{
\code{bcgpMCMC} draws samples from the Bayesian Composite Gaussian Process model
}
\details{
This draws samples from the posterior distribution for the Bayesian
Composite Gaussian Process (BCGP) model.
}
\examples{

xTrain <- matrix(runif(20, 0, 10), nrow = 10, ncol = 2)
yTrain <- xTrain[, 1] + sin(xTrain[, 2])
priors <- createPrior(noise = FALSE, d = 2)
bcgp(xTrain, yTrain, priors)
}
\seealso{
Other Major functions: \code{\link{bcgp}}
}
\concept{Major functions}
