% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrior.R
\name{createPrior}
\alias{createPrior}
\title{Create a list with prior information.}
\usage{
createPrior(deterministic = TRUE, d)
}
\arguments{
\item{deterministic}{If the data is assumed to be noise-free, then
\code{deterministic} should be \code{TRUE}. Otherwise, it should be
\code{FALSE},}

\item{d}{The dimension of the training data.}
}
\value{
A list containing the default values for all the prior parameters.
}
\description{
\code{getCorMat} returns a list that contains default values for the priors.
}
\details{
This creates a list that contains default values for the priors. The user
can change the values as they like prior to inputting the list into
\code{bcgp}.
}
\examples{
createPrior(d = 2)
createPrior(deterministic = FALSE, d = 3)
}
\concept{preprocessing functions}
