% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorMat.R
\name{getCorMat}
\alias{getCorMat}
\title{Create a correlation matrix.}
\usage{
getCorMat(x, rho)
}
\arguments{
\item{x}{An \emph{n x d} matrix.}

\item{rho}{A vector of length d.}
}
\value{
An \emph{n x n} correlation matrix
}
\description{
\code{getCorMat} returns a correlation matrix.
}
\details{
This creates a correlation matrix, \emph{R}, where \deqn{R_{ij} = \prod_{k =
1}^{d}\rho_k^{16\left( x_{ik} - x_{jk} \right)^2}}{R_ij = \prod_{k =
1}^{d}\rho_k^{16\left( x_{ik} - x_{jk} \right)^2}}
}
\examples{
n <- 10
d <- 2
x <- matrix(runif(n * d), nrow = n, ncol = d)
rho <- runif(d, 0, 1)
getCorMat(x, rho)
}
\concept{correlation and covariance functions}
