% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a vector to [0, 1]}
\usage{
rescale(x)
}
\arguments{
\item{\code{x}}{A vector input}
}
\value{
A vector rescaled to [0, 1]
}
\description{
Rescales a vector to [0, 1]
}
\details{
Rescales a vector to [0, 1].
}
\examples{
rescale(rnorm(10, 0, 100))
}
\author{
Casey Davis (\email{cbdavis33@gmail.com})
}
