% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgp.R
\name{bcgp}
\alias{bcgp}
\title{Draw samples from a bcgp model}
\usage{
bcgp(xTrain, yTrain, priors = createPrior(noise = TRUE, d =
  ncol(xTrain)), numUpdates = 5, numAdapt = 1000, burnin = 1000,
  nmcmc = 10000)
}
\arguments{
\item{xTrain}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{yTrain}{Avector containing the observed response values in the training
set.}

\item{numUpdates}{The number of updates in the proposal stepsize adaptation phase.}

\item{numAdapt}{The number of samples within each update in the proposal stepsize
adaptation phase.}

\item{burnin}{The number of burnin samples to discard after the stepsize
adaptation phase is finished}

\item{prior}{A list containing the values for the prior parameters.}
}
\value{
An object of S4 class \code{bcgp} representing the fitted results.
}
\description{
\code{bcgp} draws samples from the Bayesian Composite Gaussian Process model
}
\details{
This draws samples from the posterior distribution for the Bayesian
Composite Gaussian Process (BCGP) model.
}
\examples{

xTrain <- matrix(runif(20, 0, 10), nrow = 10, ncol = 2)
yTrain <- x1 + sin(x2)
priors <- createPrior(noise = FALSE, d = 2)
bcgp(xTrain, yTrain, priors)
}
\concept{Major functions}
