% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrior.R
\name{createPrior}
\alias{createPrior}
\title{Create a list with prior information.}
\usage{
createPrior(xTrain, noise = FALSE)
}
\arguments{
\item{xTrain}{An \code{n x d} matrix containing the independent variables
in the training set.}

\item{noise}{If the data is assumed to be noise-free, then
\code{noise} should be \code{FALSE}. Otherwise, it should be
\code{TRUE},}
}
\value{
A list containing the default values for all the prior parameters.
}
\description{
\code{createPrior} returns a list that contains default values for the priors.
}
\details{
This creates a list that contains default values for the priors. The user
can change the values as they like prior to inputting the list into
\code{bcgp}.
}
\examples{
xTrain <- matrix(runif(40), ncol= 4, nrow = 10)
createPrior(xTrain)
createPrior(xTrain, noise = TRUE)
}
\seealso{
\code{\link{bcgp}}

Other preprocessing functions: \code{\link{createInit}}
}
\concept{preprocessing functions}
